Discount codes

+++ src/main/java/pl/thinkdata/b2bbase/company/dto/SubscriptionOrderToCatalog.java
@@ -10,6 +10,7 @@ import pl.thinkdata.b2bbase.company.model.enums.SubscriptionTypeEnum;
 import pl.thinkdata.b2bbase.company.model.enums.PaymentStatusEnum;
 import pl.thinkdata.b2bbase.company.model.enums.PaymentTypeEnum;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 @Getter
@@ -22,7 +23,7 @@ public class SubscriptionOrderToCatalog {
     private SubscriptionTypeEnum packageType;
     private Date startDate;
     private Date endDate;
-    private int price;
+    private BigInteger price;
     private PaymentTypeEnum paymentType;
     private PaymentStatusEnum paymentStatus;
 }
